requirements.txt
Flask
Twilio
requests
python-dotenv
PyMuPDF
google-generativeai
faiss-cpu
